extract https://github.com/mruby/mruby @ 58fb6f421ad9824caae5a2cabb849d48647b90ec () at deps/mruby

+++ deps/mruby/mrbgems/mruby-print/src/print.c
@@ -19,7 +19,7 @@ printstr(mrb_state *mrb, mrb_value obj)
 #if defined(_WIN32)
     if (isatty(fileno(stdout))) {
       DWORD written;
-      int mlen = RSTRING_LEN(obj);
+      int mlen = (int)RSTRING_LEN(obj);
       char* utf8 = RSTRING_PTR(obj);
       int wlen = MultiByteToWideChar(CP_UTF8, 0, utf8, mlen, NULL, 0);
       wchar_t* utf16 = (wchar_t*)mrb_malloc(mrb, (wlen+1) * sizeof(wchar_t));
